minor bug fixes

+++ main.py
@@ -16,6 +16,7 @@ def is_valid_image(image_path):
 def main():
     source_folder = "~/.quickshot_temp/"
     source_folder = os.path.expanduser(source_folder)
+    print(source_folder)
     while True:
         files = os.listdir(source_folder) 
         for file in files:
@@ -24,12 +25,13 @@ def main():
             if is_valid_image(full_path):
                 image = Image.open(full_path)
                 text = extractText(full_path)
-                print(f"{file} contains text: \n{text}")
-                pyperclip.copy(text)
-                print("Text copied to clipboard")
-                os.remove(full_path)
-                print("file deleted")
-        time.sleep(1)
+                if text != "":
+                    print(f"{file} contains text: \n{text}")
+                    pyperclip.copy(text)
+                    print("Text copied to clipboard")
+                    os.remove(full_path)
+                    print("File deleted")
+        time.sleep(0.5)
 
 if __name__ == "__main__":
     main()
